style login

+++ db/migrate/20190808052000_create_condos.rb
@@ -8,12 +8,12 @@ class CreateCondos < ActiveRecord::Migration[5.0]
       t.string :address
       t.string :builder
       t.string :city
-      t.decimal :cost1bed
-      t.decimal :cost1bedden
-      t.decimal :cost2bed
-      t.decimal :cost2bedden
-      t.decimal :cost3bed
-      t.decimal :cost3bedden
+      t.integer :cost1bed
+      t.integer :cost1bedden
+      t.integer :cost2bed
+      t.integer :cost2bedden
+      t.integer :cost3bed
+      t.integer :cost3bedden
       t.integer :occupancy
